not-working search function
got the chat-gpt api to actually give results.  now we might want to do some string-manipulation to just get the parts of the response we want.  Can be difficult since the AI responses are not consistent.  It works for me because I have the AI key saved on my computer, but you'll have to add it as an environment variable to your computers for you to be able to use the API.  when we add it to digital ocean we can add the key to that server.